feat: Combine tick behaviors into single VxTickBehavior and implement NetSync filtering
